refactored LeafViewFactory
