+++ Sources/SteamPress/Views/protocol/LinkViewFactory.swift
@@ -1,0 +1,9 @@
+import Vapor
+import URI
+import Fluent
+import Foundation
+
+protocol LinkViewFactory {
+    func createLinkView(isEditing: Bool, linkToEdit: BlogLink?) throws -> View
+    func allLinksView(uri: URI, allLinks: [BlogLink]) throws -> View
+}
